[1/2] Remove accidentally uploaded class files
Class files should not be included on the repository. It's bad
practice, clutters up GitHub, and increases repostiory size.
Class files should be reserved for releases. In compliance,
this patch removes the existing class files. The next patch
of this sequence prohibits this from happening again through
adding a .gitignore which deletes all .class and .jar files.